fixes http observation response bug

+++ src/http_handler/http_response/observation.rs
@@ -5,8 +5,8 @@ pub struct ObservationResponse {
     state: String,
     angle: String,
     simulation_speed: u32,
-    width_x: f64,
-    height_y: f64,
+    width_x: u32,
+    height_y: u32,
     vx: f64,
     vy: f64,
     battery: f32,
@@ -15,10 +15,10 @@ pub struct ObservationResponse {
     distance_covered: f64,
     area_covered: AreaCoveredByLens,
     data_volume: DataVolume,
-    images_taken: u16,
-    active_time: u32,
-    objectives_done: u8,
-    objectives_points: u16,
+    images_taken: u32,
+    active_time: f64,
+    objectives_done: u16,
+    objectives_points: u32,
     timestamp: chrono::DateTime<chrono::Utc>,
 }
 
@@ -28,8 +28,8 @@ impl ObservationResponse {
     pub fn state(&self) -> &str { self.state.as_str() }
     pub fn angle(&self) -> &str { self.angle.as_str() }
     pub fn simulation_speed(&self) -> u32 { self.simulation_speed }
-    pub fn pos_x(&self) -> f64 { self.width_x }
-    pub fn pos_y(&self) -> f64 { self.height_y }
+    pub fn pos_x(&self) -> u32 { self.width_x }
+    pub fn pos_y(&self) -> u32 { self.height_y }
     pub fn vel_x(&self) -> f64 { self.vx }
     pub fn vel_y(&self) -> f64 { self.vy }
     pub fn battery(&self) -> f32 { self.battery }
